educational_website/requirements.txt
Django==4.2
opencv-python<4.7
face-recognition==1.3.0
django-crispy-forms==2.0
Pillow==9.4.0
google-generativeai
python-dotenv